requirements.txt
pyaudio
openai-whisper